[ARMEK-506] Fix Pi calculation.
Updated the 'apfloat' jar file to use the original (version 1.8.2) from
http://www.apfloat.org/apfloat_java/download.html

The temporary files needed for the calculation are stored inside
'/data/data/com.digi.android.sample.system.cpu/cache'. When the calculation
finishes all of them are automatically removed.

Used the 'setAlive(boolean)' method of class 'Pi' to cancel the process.

The maximum number of decimals is now 30000000 (long).

https://jira.digi.com/browse/ARMEK-506

Signed-off-by: Tatiana Leon <tatiana.leon@digi.com>